Rip out multisig-ism
