+++ rust/sealevel/multisig-ism/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hyperlane-sealevel-ism-multisig-ism-message-id"
+name = "multisig-ism"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,12 +8,11 @@ no-entrypoint = []
 
 [dependencies]
 borsh = "0.9"
-multisig-ism = { path = "../../../multisig-ism" }
 num-derive = "0.3"
 num-traits = "0.2"
 solana-program = "=1.14.13"
-hyperlane-core = { path = "../../../hyperlane-core" }
-hyperlane-sealevel-mailbox = { path = "../../mailbox", features = ["no-entrypoint"] }
+hyperlane-core = { path = "../hyperlane-core" }
+# hyperlane-sealevel-mailbox = { path = "../../mailbox", features = ["no-entrypoint"] }
 thiserror = "1.0"
 
 [dev-dependencies]
